requirements.txt
google-adk>=0.5.2
google-cloud-aiplatform>=1.63.0
python-dotenv
